Conform with code standards
