+++ tests/Soql/SoqlConnectionTest.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 namespace CodeliciaTest\Soql;
 
-use Codelicia\Soql\Factory\AuthorizedClientFactoryInterface;
+use Codelicia\Soql\Factory\AuthorizedClientFactory;
 use Codelicia\Soql\SoqlConnection;
 use Codelicia\Soql\SoqlStatement;
 use PHPUnit\Framework\MockObject\MockObject;
@@ -12,18 +12,18 @@ use PHPUnit\Framework\TestCase;
 
 final class SoqlConnectionTest extends TestCase
 {
-    /** @var AuthorizedClientFactoryInterface|MockObject  */
+    /** @var AuthorizedClientFactory|MockObject  */
     private $authorizedClientFactory;
 
     protected function setUp() : void
     {
-        $this->authorizedClientFactory = $this->createMock(AuthorizedClientFactoryInterface::class);
+        $this->authorizedClientFactory = $this->createMock(AuthorizedClientFactory::class);
     }
 
     /** @test */
     public function query() : void
     {
-        $sut = new SoqlConnection($this->authorizedClientFactory);
+        $sut       = new SoqlConnection($this->authorizedClientFactory);
         $statement = $sut->query('SELECT Id From Contact');
 
         self::assertInstanceOf(SoqlStatement::class, $statement);
