+++ tests/Soql/SoqlStatementTest.php
@@ -20,7 +20,7 @@ final class SoqlStatementTest extends TestCase
         self::assertSame(
             [
                 'SELECT Id FROM Contact WHERE Name = :param1 AND Surname = :param2',
-                [1 => ':param1', 2 => ':param2']
+                [1 => ':param1', 2 => ':param2'],
             ],
             $result
         );
@@ -30,8 +30,8 @@ final class SoqlStatementTest extends TestCase
     public function it_should_not_be_able_to_bind_named_values() : void
     {
         $client = new Client();
-        $sql = 'SELECT Id FROM Contact WHERE Name = :name AND Surname = :surname';
-        $sut = new Soql\SoqlStatement($client, $sql);
+        $sql    = 'SELECT Id FROM Contact WHERE Name = :name AND Surname = :surname';
+        $sut    = new Soql\SoqlStatement($client, $sql);
 
         $this->expectException(Soql\SoqlError::class);
         $this->expectExceptionMessage('SOQL does not support named parameters to queries, use question mark (?) placeholders instead');
@@ -43,8 +43,8 @@ final class SoqlStatementTest extends TestCase
     public function it_should_not_be_able_to_bind_named_params() : void
     {
         $client = new Client();
-        $sql = 'SELECT Id FROM Contact WHERE Name = :name AND Sirname = :surname';
-        $sut = new Soql\SoqlStatement($client, $sql);
+        $sql    = 'SELECT Id FROM Contact WHERE Name = :name AND Sirname = :surname';
+        $sut    = new Soql\SoqlStatement($client, $sql);
 
         $this->expectException(Soql\SoqlError::class);
         $this->expectExceptionMessage('SOQL does not support named parameters to queries, use question mark (?) placeholders instead');
@@ -57,13 +57,11 @@ final class SoqlStatementTest extends TestCase
     public function it_should_be_able_to_execute_a_query_correctly() : void
     {
         $client = new Client();
-        $sql = 'SELECT Id FROM Contact WHERE Name = :name AND Surname = :surname';
-        $sut = new Soql\SoqlStatement($client, $sql);
+        $sql    = 'SELECT Id FROM Contact WHERE Name = :name AND Surname = :surname';
+        $sut    = new Soql\SoqlStatement($client, $sql);
 
         $result = $sut->execute([':name' => 'John', 'surname' => 'Smith']);
 
         self::assertTrue($result);
     }
-
-
 }
